use width 34% for sidepanel
There is an interaction with the `display: table; content: " ";` CSS in
the `:before` and `:after` around `#main-body` and `#sidepanel`. It
looks like this pattern is called `clearfix` and is designed to fix
issues with floats.

In Safari, this CSS causes there to be a pseudo-element with a width of `1%`,
which means that if the sum of widths of the elements adds to `100%`,
the rightmost float will drop to the next line.

I don't know if removing the `clearfix` would cause more
problems on other browsers. To avoid potentially causing even more issues
with a bugfix, this commit just changes the width of the `#main-body` and
`#sidepanel` elements to add to `99%` instead of `100%`. It will look
slightly worse on non-Safari browsers, but will no longer be broken on
Safari.